requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.3
idna==3.2
python-dotenv==0.18.0
requests==2.26.0
simple-chalk==0.1.0
urllib3==1.26.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
